feat: remove percentage progress bar
